Added 16 player functionality
Added 16 player functionality. Actually the game supports any number of players but the prizes were calculated for up to 16 possible winners.

The game has been tested with a game from start to finish, where winner account balances were verified to be correct. It is still necessary to test the contract with two consecutive games.